+++ R/lasfilter.r
@@ -75,7 +75,7 @@ lasfilter_ <- function(.las, conditions)
     return(NULL)
   }
 
-  return(LAS(.las@data[combined_bools], .las@header, .las@crs))
+  return(LAS(.las@data[combined_bools], .las@header, .las@proj4string))
 }
 
 #' Predefined filters