+++ R/lasfilterdecimate.r
@@ -129,7 +129,7 @@ lasfilterdecimate.LAS = function(las, density, homogenize = TRUE, res = 5, use_p
       selected <- las@data[, .I[selected_pulses(1:.N, n)], by = by]$V1
   }
 
-  return(LAS(las@data[selected], las@header, las@crs))
+  return(LAS(las@data[selected], las@header, las@proj4string))
 }
 
 #' @export