+++ R/lasfiltersurfacepoints.r
@@ -37,7 +37,7 @@ lasfiltersurfacepoints.LAS = function(las, res)
   Z   <- NULL
   by  <- group_grid(las@data$X, las@data$Y, res)
   sub <- las@data[las@data[, .I[which.max(Z)], by = by]$V1]
-  las <- LAS(sub, las@header, las@crs)
+  las <- LAS(sub, las@header, las@proj4string)
   return(las)
 }