+++ R/utils_misc.r
@@ -48,7 +48,7 @@ make_overlay_raster = function(las, res, start = c(0,0), subcircle = 0)
   bbox@xmax <- round_any(bbox@xmax - 0.5 * res - start[1], res) + res + start[1]
   bbox@ymin <- round_any(bbox@ymin - 0.5 * res - start[2], res) + start[2]
   bbox@ymax <- round_any(bbox@ymax - 0.5 * res - start[2], res) + res + start[2]
-  layout    <- suppressWarnings(raster::raster(bbox, res = res, crs = las@crs))
+  layout    <- suppressWarnings(raster::raster(bbox, res = res, crs = las@proj4string))
   return(layout)
 }